+++ packages/ui-helpers/src/__tests__/Footer.spec.tsx
@@ -1,3 +1,5 @@
+/* eslint-disable */
+import { describe, expect, it } from "vitest";
 // import mixpanel from "mixpanel-browser";
 // import { Footer } from "../Footer";
 // import { render, renderHook, act } from "@testing-library/react";